pyproject.toml
[project]
name = "ruby"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-speech>=2.35.0",
    "google-cloud-texttospeech>=2.33.0",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "pygame>=2.6.1",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
    "yt-dlp>=2025.12.8",
    "youtube-dl>=2021.12.17",
    "faiss-cpu>=1.13.2",
    "langchain-community>=0.4.1",
    "pyserial>=3.5",
    "pypdf>=6.6.0",
    "langchain-text-splitters>=1.1.0",
]
